requirements.txt
absl-py==2.1.0
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
argcomplete==2.0.0
attrs==19.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cmake==3.28.1
distlib==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.3.0
exceptiongroup==1.2.0
filelock==3.13.1
flake8==7.1.0
flake8-blind-except==0.2.1
flake8-builtins==2.0.0
flake8-class-newline==1.6.0
flake8-comprehensions==3.10.0
flake8-deprecated==2.0.1
flake8-docstrings==1.6.0
flake8-import-order==0.18.1
flake8-quotes==3.3.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
gitpython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.0.0
grpcio==1.62.1
gym==0.23.1
gym-notices==0.0.8
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.24.6
hydra-core==1.3.2
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==1.1.1
lit==17.0.6
loguru==0.7.0
markdown==3.5.2
markupsafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.0.7
networkx==3.1
ninja==1.11.1.1
numpy==1.22.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.0.3
pathtools==0.1.2
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
promise==2.3
protobuf==4.25.3
psutil==5.9.8
py==1.11.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycodestyle==2.12.0
pydantic==2.6.0
pydantic-core==2.16.1
pydocstyle==6.3.0
pyflakes==3.2.0
pytest==7.1.3
pytest-repeat==0.9.1
pytest-rerunfailures==10.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyvirtualdisplay==3.0
pyyaml==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.4
scipy==1.10.1
sentry-sdk==1.40.0
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
termcolor==2.3.0
timm==1.0.9
tomli==2.0.1
torch==1.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.66.1
trimesh==3.22.0
triton==2.0.0
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.0
virtualenv==20.25.0
wandb==0.16.5
werkzeug==3.0.1
zipp==3.18.1
